+++ src/index.ts
@@ -1,8 +1,9 @@
 import { Plugin } from "@yarnpkg/core";
 import { LicensesListCommand } from "./commands/list";
+import { LicensesGenerateDisclaimerCommand } from "./commands/generateDisclaimer";
 
 const plugin: Plugin = {
-  commands: [LicensesListCommand],
+  commands: [LicensesListCommand, LicensesGenerateDisclaimerCommand],
 };
 
 export default plugin;