+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.3.0] - 2021-01-28
+[0.3.0]: https://github.com/mhassan1/yarn-plugin-licenses/compare/v0.2.1...v0.3.0
+
+- Normalize expected output line endings in tests
+- Exclude duplicated virtual/non-virtual descriptors from output
+- Add `generate-disclaimer` command
+- Refactor `getPackageManifest` into `getPackagePath` in linkers
+- Compress paths in `ppath.join`
+
 ## [0.2.1] - 2020-12-30
 [0.2.1]: https://github.com/mhassan1/yarn-plugin-licenses/compare/v0.2.0...v0.2.1